requirements.txt
Flask
openai
python-dotenv
apscheduler